Move nano::block_sideband and nano::block_type in to their own files.
